fractional moose I guess...

+++ Population.java
@@ -17,7 +17,7 @@ public class Population {
         currentMoosePop -= newMoose;
         currentMoosePop += deadMoose;
         
-        if((int)currentMoosePop >= 1) return 1 + getInitialStartTimeForMoose1((int)currentMoosePop);
+        if((int)currentMoosePop >= 1) return 1 + getInitialStartTimeForMoose1(currentMoosePop);
         return 0;
     }
     public static int getInitialStartTimeForMoose2(double currentMoosePop, double startingMoosePopulation) {
